Fix compiler errors.

+++ src/lib/storage/chunk.cpp
@@ -160,7 +160,8 @@ const PolymorphicAllocator<Chunk>& Chunk::get_allocator() const { return _alloc;
 uint64_t Chunk::AccessCounter::history_sample(size_t lookback) const {
   if (_history.size() < 2 || lookback == 0) return 0;
   const auto last = _history.back();
-  const auto prelast_index = std::max(0ll, static_cast<int64_t>(_history.size()) - static_cast<int64_t>(lookback));
+  const auto prelast_index =
+      std::max(static_cast<int64_t>(0), static_cast<int64_t>(_history.size()) - static_cast<int64_t>(lookback));
   const auto prelast = _history.at(prelast_index);
   return last - prelast;
 }
